refactor: LottosFactory 패키지 이동
