+++ src/test/java/lotto/controller/LottosFactoryTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.controller;
 
 import static org.assertj.core.api.Assertions.*;
 import org.junit.jupiter.api.DisplayName;
